Remove Sha1Sum in favor of hex_to_sha / sha_to_hex.
The Sha1Sum code adds object creation overhead everywhere, which slows Dulwich
down (further), as well as memory overhead though I doubt in a significant
manner.

Sha1Sum is quite convenient when programming inside of Dulwich itself, but less
useful for users of Dulwich: they have to create and pass in a custom object,
and have to deal with a custom object.